NLP/AdvancedTopics/nlp-project/requirements.txt
nltk
spacy
scikit-learn
transformers
openai
flask
python-dotenv
